requirements.txt
Django==4.2.11
Pillow==11.0.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
